Connections
Connected the database with the html pages in the front end.
pages now able to fetch images from the DB as well.
connected the nav links of the pages..
still some coding left in the php section .
Styling left as well, need to improve the consistency in the website as a whole.